Redesigned memory cache, removed impls
