+++ library/src/com/nostra13/universalimageloader/cache/memory/impl/LruMemoryCache.java
@@ -1,13 +1,12 @@
 package com.nostra13.universalimageloader.cache.memory.impl;
 
 import android.graphics.Bitmap;
-
 import com.nostra13.universalimageloader.cache.memory.MemoryCache;
 
-import java.util.Collection;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * A cache that holds strong references to a limited number of Bitmaps. Each time a Bitmap is accessed, it is moved to
@@ -53,7 +52,7 @@ public class LruMemoryCache implements MemoryCache {
 
 	/** Caches {@code Bitmap} for {@code key}. The Bitmap is moved to the head of the queue. */
 	@Override
-	public final boolean put(String key, Bitmap value) {
+	public final void put(String key, Bitmap value) {
 		if (key == null || value == null) {
 			throw new NullPointerException("key == null || value == null");
 		}
@@ -67,7 +66,11 @@ public class LruMemoryCache implements MemoryCache {
 		}
 
 		trimToSize(maxSize);
-		return true;
+	}
+
+	@Override
+	public int size() {
+		return size;
 	}
 
 	/**
@@ -117,7 +120,7 @@ public class LruMemoryCache implements MemoryCache {
 	}
 
 	@Override
-	public Collection<String> keys() {
+	public Set<String> keys() {
 		synchronized (this) {
 			return new HashSet<String>(map.keySet());
 		}
